refactor: move regex capture helpers to nginx-lint-plugin
`find_unnamed_capture_positions` and `regex_has_named_capture` were private
to the nginx_rift plugin, but the same PCRE parsing is needed by any rule
that reasons about capture groups. Move them to
`nginx_lint_plugin::helpers` with doctests, alongside a `has_unnamed_capture`
convenience wrapper.

No behaviour change: nginx_rift now imports what it used to define.
